requirements.txt
matplotlib==3.5.2
pandas==1.4.1
Pillow==9.1.1
plotly==5.8.2
psutil==5.8.0
PyOpenGL==3.1.6
PyQt5==5.15.7
rdkit==2022.3.3
scikit_learn==1.1.1
scipy==1.8.1
tensorflow==2.9
